pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "payday-mvp"
version = "0.1.0"
description = "Streamlit MVP for payday call upload, transcription, analysis, and dashboarding."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.43.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
